// src/types.ts
export interface ProjectItem {
  id: number;
  title: string;
  text: string;
  link: string;
  shortLabel: string;
}

export interface WhyCardItem {
  number: string;
  title: string;
  description: string;
}

export interface ProcessItem {
  step: string;
  title: string;
  description: string;
}